requirements.txt
numpy==1.16.0
python-Levenshtein==0.12.0
scikit-learn==0.20.3
scipy==1.2.1
